Scala API
