--- scala/api/src/main/scala/com/tersesystems/echopraxia/sapi/support/AbstractLoggerSupport.scala
@@ -1,8 +1,0 @@
-package com.tersesystems.echopraxia.sapi.support
-
-import com.tersesystems.echopraxia.core.CoreLogger
-
-abstract class AbstractLoggerSupport[FB](val core: CoreLogger, val fieldBuilder: FB)
-    extends DefaultMethodsSupport[FB] {
-  def name: String = core.getName
-}
