--- scala/api/src/main/scala/com/tersesystems/echopraxia/sapi/support/DefaultMethodsSupport.scala
@@ -1,11 +1,0 @@
-package com.tersesystems.echopraxia.sapi.support
-
-import com.tersesystems.echopraxia.core.CoreLogger
-
-trait DefaultMethodsSupport[FB] {
-  def name: String
-
-  def core: CoreLogger
-
-  def fieldBuilder: FB
-}
