+++ scala/api/src/main/scala/com/tersesystems/echopraxia/scala/api/DefaultMethodsSupport.scala
@@ -1,0 +1,11 @@
+package com.tersesystems.echopraxia.scala.api
+
+import com.tersesystems.echopraxia.api.CoreLogger
+
+trait DefaultMethodsSupport[FB] {
+  def name: String
+
+  def core: CoreLogger
+
+  def fieldBuilder: FB
+}
