--- scala/api/src/test/resources/logback-test.xml
@@ -1,10 +1,0 @@
-<configuration>
-    <appender name="LIST" class="ch.qos.logback.core.read.ListAppender">
-    </appender>
-
-    <logger name="com.tersesystems.echopraxia.sapi" level="TRACE"/>
-
-    <root level="INFO">
-        <appender-ref ref="LIST"/>
-    </root>
-</configuration>
