--- scala/sourcecode/src/main/scala/com/tersesystems/echopraxia/sapi/sourcecode/support/DefaultMethodsSupport.scala
@@ -1,7 +1,0 @@
-package com.tersesystems.echopraxia.sapi.sourcecode.support
-
-/**
-0 * DefaultMethodsSupport implementation (alias to sapi.support)
- */
-trait DefaultMethodsSupport[FB]
-    extends com.tersesystems.echopraxia.sapi.support.DefaultMethodsSupport[FB]
