--- scala/sourcecode/src/main/scala/com/tersesystems/echopraxia/sapi/sourcecode/support/LoggerSupport.scala
@@ -1,6 +1,0 @@
-package com.tersesystems.echopraxia.sapi.sourcecode.support
-
-/**
- * Logger support methods (an alias to sapi.support.LoggerSupport)
- */
-trait LoggerSupport[FB] extends com.tersesystems.echopraxia.sapi.support.LoggerSupport[FB]
