+++ scala/sourcecode/src/main/scala/com/tersesystems/echopraxia/scala/LoggerMethods.scala
@@ -1,0 +1,353 @@
+package com.tersesystems.echopraxia.scala
+
+import com.tersesystems.echopraxia.api.FieldBuilderResult
+import com.tersesystems.echopraxia.scala.api.Condition
+
+/**
+ * Logger methods with source code implicits
+ */
+trait LoggerMethods[FB] {
+
+  /** @return true if the logger level is TRACE or higher. */
+  def isTraceEnabled: Boolean
+
+  /**
+   * @param condition the given condition.
+   * @return true if the logger level is TRACE or higher and the condition is met.
+   */
+  def isTraceEnabled(condition: Condition): Boolean
+
+  /**
+   * Logs statement at TRACE level.
+   *
+   * @param message the given message.
+   */
+  def trace(
+             message: String
+           )(implicit line: sourcecode.Line, file: sourcecode.File, enc: sourcecode.Enclosing): Unit
+
+  /**
+   * Logs statement at TRACE level using a field builder function.
+   *
+   * @param message the message.
+   * @param f       the field builder function.
+   */
+  def trace(message: String, f: FB => FieldBuilderResult)(implicit
+                                                             line: sourcecode.Line,
+                                                             file: sourcecode.File,
+                                                             enc: sourcecode.Enclosing
+  ): Unit
+
+  /**
+   * Logs statement at TRACE level with exception.
+   *
+   * @param message the message.
+   * @param e       the given exception.
+   */
+  def trace(message: String, e: Throwable)(implicit
+                                           line: sourcecode.Line,
+                                           file: sourcecode.File,
+                                           enc: sourcecode.Enclosing
+  ): Unit
+
+  /**
+   * Conditionally logs statement at TRACE level.
+   *
+   * @param condition the given condition.
+   * @param message   the message.
+   */
+  def trace(condition: Condition, message: String)(implicit
+                                                   line: sourcecode.Line,
+                                                   file: sourcecode.File,
+                                                   enc: sourcecode.Enclosing
+  ): Unit
+
+  /**
+   * Conditionally logs statement at TRACE level using a field builder function.
+   *
+   * @param condition the given condition.
+   * @param message   the message.
+   * @param f         the field builder function.
+   */
+  def trace(condition: Condition, message: String, f: FB => FieldBuilderResult)(implicit
+                                                                                   line: sourcecode.Line,
+                                                                                   file: sourcecode.File,
+                                                                                   enc: sourcecode.Enclosing
+  ): Unit
+
+  /**
+   * Conditionally logs statement at TRACE level with exception.
+   *
+   * @param condition the given condition.
+   * @param message   the message.
+   * @param e         the given exception.
+   */
+  def trace(condition: Condition, message: String, e: Throwable)(implicit
+                                                                 line: sourcecode.Line,
+                                                                 file: sourcecode.File,
+                                                                 enc: sourcecode.Enclosing
+  ): Unit
+
+  /** @return true if the logger level is DEBUG or higher. */
+  def isDebugEnabled: Boolean
+
+  /**
+   * @param condition the given condition.
+   * @return true if the logger level is DEBUG or higher and the condition is met.
+   */
+  def isDebugEnabled(condition: Condition): Boolean
+
+  /**
+   * Logs statement at DEBUG level.
+   *
+   * @param message the given message.
+   */
+  def debug(
+             message: String
+           )(implicit line: sourcecode.Line, file: sourcecode.File, enc: sourcecode.Enclosing): Unit
+
+  /**
+   * Logs statement at DEBUG level using a field builder function.
+   *
+   * @param message the message.
+   * @param f       the field builder function.
+   */
+  def debug(message: String, f: FB => FieldBuilderResult)(implicit
+                                                             line: sourcecode.Line,
+                                                             file: sourcecode.File,
+                                                             enc: sourcecode.Enclosing
+  ): Unit
+
+  /**
+   * Logs statement at DEBUG level with exception.
+   *
+   * @param message the message.
+   * @param e       the given exception.
+   */
+  def debug(message: String, e: Throwable)(implicit
+                                           line: sourcecode.Line,
+                                           file: sourcecode.File,
+                                           enc: sourcecode.Enclosing
+  ): Unit
+
+  /**
+   * Conditionally logs statement at DEBUG level.
+   *
+   * @param condition the given condition.
+   * @param message   the message.
+   */
+  def debug(condition: Condition, message: String)(implicit
+                                                   line: sourcecode.Line,
+                                                   file: sourcecode.File,
+                                                   enc: sourcecode.Enclosing
+  ): Unit
+
+  /**
+   * Conditionally logs statement at DEBUG level with exception.
+   *
+   * @param condition the given condition.
+   * @param message   the message.
+   * @param e         the given exception.
+   */
+  def debug(condition: Condition, message: String, e: Throwable)(implicit
+                                                                 line: sourcecode.Line,
+                                                                 file: sourcecode.File,
+                                                                 enc: sourcecode.Enclosing
+  ): Unit
+
+  /**
+   * Conditionally logs statement at DEBUG level using a field builder function.
+   *
+   * @param condition the given condition.
+   * @param message   the message.
+   * @param f         the field builder function.
+   */
+  def debug(condition: Condition, message: String, f: FB => FieldBuilderResult)(implicit
+                                                                                   line: sourcecode.Line,
+                                                                                   file: sourcecode.File,
+                                                                                   enc: sourcecode.Enclosing
+  ): Unit
+
+  /** @return true if the logger level is INFO or higher. */
+  def isInfoEnabled: Boolean
+
+  /**
+   * @param condition the given condition.
+   * @return true if the logger level is INFO or higher and the condition is met.
+   */
+  def isInfoEnabled(condition: Condition): Boolean
+
+  /**
+   * Logs statement at INFO level.
+   *
+   * @param message the given message.
+   */
+  def info(
+            message: String
+          )(implicit line: sourcecode.Line, file: sourcecode.File, enc: sourcecode.Enclosing): Unit
+
+  /**
+   * Logs statement at INFO level using a field builder function.
+   *
+   * @param message the message.
+   * @param f       the field builder function.
+   */
+  def info(message: String, f: FB => FieldBuilderResult)(implicit
+                                                            line: sourcecode.Line,
+                                                            file: sourcecode.File,
+                                                            enc: sourcecode.Enclosing
+  ): Unit
+
+  /**
+   * Logs statement at INFO level with exception.
+   *
+   * @param message the message.
+   * @param e       the given exception.
+   */
+  def info(message: String, e: Throwable)(implicit
+                                          line: sourcecode.Line,
+                                          file: sourcecode.File,
+                                          enc: sourcecode.Enclosing
+  ): Unit
+
+  /**
+   * Conditionally logs statement at INFO level.
+   *
+   * @param condition the given condition.
+   * @param message   the message.
+   */
+  def info(condition: Condition, message: String)(implicit
+                                                  line: sourcecode.Line,
+                                                  file: sourcecode.File,
+                                                  enc: sourcecode.Enclosing
+  ): Unit
+
+  /**
+   * Conditionally logs statement at INFO level using a field builder function.
+   *
+   * @param condition the given condition.
+   * @param message   the message.
+   * @param f         the field builder function.
+   */
+  def info(condition: Condition, message: String, f: FB => FieldBuilderResult)(implicit
+                                                                                  line: sourcecode.Line,
+                                                                                  file: sourcecode.File,
+                                                                                  enc: sourcecode.Enclosing
+  ): Unit
+
+  /**
+   * Conditionally logs statement at INFO level with exception.
+   *
+   * @param condition the given condition.
+   * @param message   the message.
+   * @param e         the given exception.
+   */
+  def info(condition: Condition, message: String, e: Throwable)(implicit
+                                                                line: sourcecode.Line,
+                                                                file: sourcecode.File,
+                                                                enc: sourcecode.Enclosing
+  ): Unit
+
+  /** @return true if the logger level is WARN or higher. */
+  def isWarnEnabled: Boolean
+
+  /**
+   * @param condition the given condition.
+   * @return true if the logger level is WARN or higher and the condition is met.
+   */
+  def isWarnEnabled(condition: Condition): Boolean
+
+  /**
+   * Logs statement at WARN level.
+   *
+   * @param message the given message.
+   */
+  def warn(
+            message: String
+          )(implicit line: sourcecode.Line, file: sourcecode.File, enc: sourcecode.Enclosing): Unit
+
+  /**
+   * Logs statement at WARN level using a field builder function.
+   *
+   * @param message the message.
+   * @param f       the field builder function.
+   */
+  def warn(message: String, f: FB => FieldBuilderResult)(implicit
+                                                            line: sourcecode.Line,
+                                                            file: sourcecode.File,
+                                                            enc: sourcecode.Enclosing
+  ): Unit
+
+  /**
+   * Logs statement at WARN level with exception.
+   *
+   * @param message the message.
+   * @param e       the given exception.
+   */
+  def warn(message: String, e: Throwable)(implicit
+                                          line: sourcecode.Line,
+                                          file: sourcecode.File,
+                                          enc: sourcecode.Enclosing
+  ): Unit
+
+  def warn(condition: Condition, message: String)(implicit
+                                                  line: sourcecode.Line,
+                                                  file: sourcecode.File,
+                                                  enc: sourcecode.Enclosing
+  ): Unit
+
+  def warn(condition: Condition, message: String, e: Throwable)(implicit
+                                                                line: sourcecode.Line,
+                                                                file: sourcecode.File,
+                                                                enc: sourcecode.Enclosing
+  ): Unit
+
+  def warn(condition: Condition, message: String, f: FB => FieldBuilderResult)(implicit
+                                                                                  line: sourcecode.Line,
+                                                                                  file: sourcecode.File,
+                                                                                  enc: sourcecode.Enclosing
+  ): Unit
+
+  /** @return true if the logger level is ERROR or higher. */
+  def isErrorEnabled: Boolean
+
+  /**
+   * @param condition the given condition.
+   * @return true if the logger level is ERROR or higher and the condition is met.
+   */
+  def isErrorEnabled(condition: Condition): Boolean
+
+  def error(
+             message: String
+           )(implicit line: sourcecode.Line, file: sourcecode.File, enc: sourcecode.Enclosing): Unit
+
+  def error(message: String, f: FB => FieldBuilderResult)(implicit
+                                                             line: sourcecode.Line,
+                                                             file: sourcecode.File,
+                                                             enc: sourcecode.Enclosing
+  ): Unit
+
+  def error(message: String, e: Throwable)(implicit
+                                           line: sourcecode.Line,
+                                           file: sourcecode.File,
+                                           enc: sourcecode.Enclosing
+  ): Unit
+
+  def error(condition: Condition, message: String)(implicit
+                                                   line: sourcecode.Line,
+                                                   file: sourcecode.File,
+                                                   enc: sourcecode.Enclosing
+  ): Unit
+
+  def error(condition: Condition, message: String, f: FB => FieldBuilderResult)(implicit
+                                                                                   line: sourcecode.Line,
+                                                                                   file: sourcecode.File,
+                                                                                   enc: sourcecode.Enclosing
+  ): Unit
+
+  def error(condition: Condition, message: String, e: Throwable)(implicit
+                                                                 line: sourcecode.Line,
+                                                                 file: sourcecode.File,
+                                                                 enc: sourcecode.Enclosing
+  ): Unit
+}
