+++ scala/version.sbt
@@ -1,1 +1,1 @@
-ThisBuild / version := "1.5.0-SNAPSHOT"
+ThisBuild / version := "2.0.0-SNAPSHOT"
